+++ sw.js
@@ -1,5 +1,4 @@
 /* eslint-env browser, serviceworker */
-
 'use strict';
 
 /* globals getJSON, activateCallback, WorkInfo */
@@ -10,7 +9,8 @@
 // import getJSON from './node_modules/simple-get-json/dist/index-es.js';
 // import activateCallback from 'node_modules/textbrowser/resources/activateCallback.js';
 // import {getWorkFiles} from './WorkInfo.js';
-importScripts('node_modules/@babel/polyfill/dist/polyfill.js');
+importScripts('node_modules/core-js-bundle/minified.js');
+importScripts('node_modules/regenerator-runtime/runtime.js');
 importScripts('node_modules/simple-get-json/dist/index.js');
 importScripts('node_modules/textbrowser/dist/WorkInfo-umd.js');
 importScripts('node_modules/textbrowser/dist/activateCallback-umd.js');
@@ -129,7 +129,8 @@ const defaultUserStaticFiles = [
 //   get these as reliable full paths without hard-coding or needing to
 //   actually be in `node_modules/textbrowser`; see `resources/index.js`
 const textbrowserStaticResourceFiles = [
-    'node_modules/@babel/polyfill/dist/polyfill.js',
+    'node_modules/core-js-bundle/minified.js',
+    'node_modules/regenerator-runtime/runtime.js',
     'node_modules/dialog-polyfill/dist/dialog-polyfill.esm.js',
 
     'node_modules/textbrowser/appdata/languages.json',